requirements.txt
appdirs==1.4.4
asgiref==3.2.10
beautifulsoup4==4.9.3
certifi==2020.12.5
chardet==4.0.0
distlib==0.3.1
Django==3.1.2
django-bootstrap-v5==1.0.0
django-bootstrap4==2.3.1
django-crispy-forms==1.10.0
django-embed-video==1.4.0
filelock==3.0.12
idna==2.10
Jinja2==2.11.2
MarkupSafe==1.1.1
Pillow==8.0.1
psycopg2-binary==2.8.6
pytz==2020.1
requests==2.25.1
six==1.15.0
soupsieve==2.1
sqlparse==0.4.1
urllib3==1.26.3
